@@ -97,8 +97,7 @@ return 0;
     elif os.path.basename(cc[0]).find('clang') != -1:
         ccmd.extend( ['-L/opt/local/lib/libomp', '-lomp'] )
     try:
-        with open('/dev/null', 'wb') as devnull:
-            output = subprocess.check_call(ccmd, stderr=devnull)
+        output = subprocess.check_call(ccmd, stderr=subprocess.DEVNULL)
         outCFLAGS = ['-fopenmp',]
         outLIBS = []
         if os.path.basename(cc[0]).find('gcc') != -1:
@@ -107,7 +106,7 @@ return 0;
             outLIBS.extend( ['-L/opt/local/lib/libomp', '-lomp'] )
             
     except subprocess.CalledProcessError:
-        print("WARNING:  OpenMP does not appear to be supported by %s, disabling" % cc[0])
+        print(f"WARNING:  OpenMP does not appear to be supported by {cc[0]}, disabling")
         outCFLAGS = []
         outLIBS = []
         
@@ -159,7 +158,7 @@ def get_fftw():
             pass
             
         if len(outVersion) > 0:
-            print("Found FFTW3, version %s" % str(outVersion[0]))
+            print(f"Found FFTW3, version {outVersion[0]}")
             
     except (OSError, subprocess.CalledProcessError):
         print("WARNING:  single precision FFTW3 cannot be found, using defaults")
@@ -175,8 +174,7 @@ def write_version_info():
     lslVersion = get_version()
     shortVersion = '.'.join(lslVersion.split('.')[:2])
     
-    contents = """# -*- coding: utf-8 -*-
-# This file is automatically generated by setup.py
+    contents = f"""# This file is automatically generated by setup.py
 
 import os
 import glob
@@ -209,11 +207,11 @@ def get_fingerprint():
         m.update(_get_md5(filename))
     return m.hexdigest()
 
-version = '%s'
-full_version = '%s'
-short_version = '%s'
+version = '{lslVersion}'
+full_version = '{lslVersion}'
+short_version = '{shortVersion}'
 
-""" % (lslVersion, lslVersion, shortVersion)
+"""
     
     with open('lsl/version/__init__.py', 'w') as fh:
         fh.write(contents)
@@ -265,8 +263,8 @@ class lsl_build_ext(build_ext):
         if os.getenv('WITH_FFTW', None) is not None:
             self.with_fftw = os.getenv('WITH_FFTW')
         if self.with_fftw is not None:
-            fftwFlags = ['-I%s/include' % self.with_fftw,]
-            fftwLibs = ['-L%s/lib' % self.with_fftw, '-lfftw3f']
+            fftwFlags = [f"-I{self.with_fftw}/include",]
+            fftwLibs = [f"-L{self.with_fftw}/lib", '-lfftw3f']
         else:
             fftwFlags, fftwLibs = get_fftw()
             
@@ -326,18 +324,19 @@ setup(
                    'Intended Audience :: Science/Research',
                    'License :: OSI Approved :: GNU General Public License (GPL)',
                    'Topic :: Scientific/Engineering :: Astronomy',
-                   'Programming Language :: Python :: 2',
-                   'Programming Language :: Python :: 2.7',
                    'Programming Language :: Python :: 3',
-                   'Programming Language :: Python :: 3.5',
                    'Programming Language :: Python :: 3.6',
                    'Programming Language :: Python :: 3.7',
                    'Programming Language :: Python :: 3.8',
+                   'Programming Language :: Python :: 3.9',
+                   'Programming Language :: Python :: 3.10',
+                   'Programming Language :: Python :: 3.11',
+                   'Programming Language :: Python :: 3.12',
                    'Operating System :: MacOS :: MacOS X',
                    'Operating System :: POSIX :: Linux'],
     packages = find_namespace_packages(), 
     scripts = glob.glob('scripts/*.py'), 
-    python_requires='>=2.7', 
+    python_requires='>=3.6', 
     setup_requires = ['numpy>=1.7'], 
     install_requires = ['astropy>=3.0', 'healpy', 'numpy>=1.7', 'scipy>=0.19', 'pyephem>=3.7.5.3', 'aipy>=3.0.1', 'pytz>=2012c'],
     include_package_data = True,  
