@@ -334,9 +334,9 @@ def _self_cal(aa, dataSet, simSet, chan, pol, ref_ant=0, max_iter=30, amplitude=
 
     # Make sure we have the right polarization
     if pol not in dataSet.pols:
-        raise RuntimeError("Data set does not have data for polarization '%s'" % pol)
+        raise RuntimeError(f"Data set does not have data for polarization '{pol}'")
     if pol not in simSet.pols:
-        raise RuntimeError("Simulation set does not have data for polarization '%s'" % pol)
+        raise RuntimeError(f"Simulation set does not have data for polarization '{pol}'")
 
     # Make sure that `chan' is an array by trying to find its length
     try:
@@ -360,9 +360,9 @@ def _self_cal(aa, dataSet, simSet, chan, pol, ref_ant=0, max_iter=30, amplitude=
     else:
         found = True
     if not found:
-        raise RuntimeError("Stand #%i not found in the array provided" % ref_ant)
+        raise RuntimeError(f"Stand #{ref_ant} not found in the array provided")
     if verbose:
-        print("Using antenna #%i as a reference (Stand #%i)" % (ref_ant, aa.ants[ref_ant].stand))
+        print(f"Using antenna #{ref_ant} as a reference (Stand #{aa.ants[ref_ant].stand})")
         
     # Frequency in GHz so that the delays can be in ns
     fq = dataSet.freq[chan] / 1e9
