@@ -228,8 +228,8 @@ class ProjectedSkyMap(object):
         
         self.skymap_object = skymap_object
         
-        assert  -90 <= lat <=  90, ValueError('lat = %g not in [-90,90]' % lat)
-        assert -360 <= lon <= 360, ValueError('lon = %g not in [-360,360]' % lon)
+        assert  -90 <= lat <=  90, ValueError(f"lat = {lat:g} not in [-90,90]")
+        assert -360 <= lon <= 360, ValueError(f"lon = {lon:g} not in [-360,360]")
         self.lat  = lat
         self.lon  = lon
         self.time = utc_jd
